@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class Main(AppConfig):
-    name = 'main'
+    name = 'server.apps.main'
