@@ -1,10 +1,10 @@
-from pathlib import PurePath
+from pathlib import Path
 
 from decouple import AutoConfig
 
 # Build paths inside the project like this: BASE_DIR.joinpath('some')
 # `pathlib` is better than writing: dirname(dirname(dirname(__file__)))
-BASE_DIR = PurePath(__file__).parent.parent.parent.parent
+BASE_DIR = Path(__file__).parent.parent.parent.parent
 
 # Loading `.env` files
 # See docs: https://gitlab.com/mkleehammer/autoconfig
